# backend/src/utils/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: TAUREAN API Documentation
  version: 1.0.0
  description: API documentation for TAUREAN Inventory Management System
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/v1/deletions/company:
    post:
      summary: Queue company hard deletion (30 days)
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companyId]
              properties:
                companyId: { type: string }
                reason: { type: string }
      responses:
        "201": { description: Queued }
...
  /api/v1/facilities/company:
    get:
      summary: Get company-specific facilities
      tags: [Facilities]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          description: Maximum number of facilities to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of facilities to skip
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, maintenance]
          description: Filter by facility status
      responses:
        "200":
          description: Company facilities retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
  /api/v1/bookings/company:
    get:
      summary: Get company-specific bookings
      tags: [Bookings]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          description: Maximum number of bookings to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of bookings to skip
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
          description: Filter by booking status
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: Filter by start date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: Filter by end date
      responses:
        "200":
          description: Company bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
  /api/v1/users/company:
    get:
      summary: Get company-specific users
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          description: Maximum number of users to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of users to skip
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, staff, user]
          description: Filter by user role
        - in: query
          name: search
          schema:
            type: string
          description: Search by name, email, or username
      responses:
        "200":
          description: Company users retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
  /api/v1/inventory-items/company:
    get:
      summary: Get company-specific inventory items
      tags: [Inventory]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          description: Maximum number of items to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of items to skip
        - in: query
          name: status
          schema:
            type: string
            enum: [available, rented, maintenance, retired]
          description: Filter by item status
        - in: query
          name: category
          schema:
            type: string
          description: Filter by item category
        - in: query
          name: search
          schema:
            type: string
          description: Search by name or description
      responses:
        "200":
          description: Company inventory items retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
  /api/v1/taxes/company:
    get:
      summary: Get company-specific taxes
      tags: [Taxes]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          description: Maximum number of taxes to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of taxes to skip
        - in: query
          name: active
          schema:
            type: boolean
          description: Filter by active status
        - in: query
          name: type
          schema:
            type: string
            enum: [percentage, fixed]
          description: Filter by tax type
      responses:
        "200":
          description: Company taxes retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"

  # Missing Invoice Routes
  /api/v1/invoices/company:
    get:
      summary: Get company-specific invoices
      tags: [Invoices]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          description: Maximum number of invoices to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of invoices to skip
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, paid, overdue, cancelled]
          description: Filter by invoice status
      responses:
        "200":
          description: Company invoices retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
...
  /api/v1/transaction/company:
    get:
      summary: Get company-specific transactions
      tags: [Transactions]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          description: Maximum number of transactions to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of transactions to skip
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, successful, failed, cancelled]
          description: Filter by transaction status
        - in: query
          name: type
          schema:
            type: string
            enum: [payment, refund, transfer]
          description: Filter by transaction type
      responses:
        "200":
          description: Company transactions retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...